// api/src/model/BackgroundProcess.entity.generated.ts
import { ApiProperty } from '@nestjs/swagger';
import { AllUser } from './AllUser.entity.generated';

/**
 * Informational model in typescript (nest.js) for BackgroundProcess
 * IMPORTANT, DO NOT EDIT: this is a generated file and is over written by subsequent runs of the generator. Any changes will be lost
 * one job that needs to be run in the background
 */

/**
 * BackgroundProcess
 * one job that needs to be run in the background
 */
export class BackgroundProcess {
	/**
	 * the unique internal identifier (primary key) of the row
	 */	
	@ApiProperty({ description: 'the unique internal identifier (primary key) of the row'}) 
	id: number;

	/**
	 * 
	 */	
	@ApiProperty({ description: ''}) 
	cmd_line: string;

	/**
	 * if set, the point in time the process is known to have started
	 */	
	@ApiProperty({ description: 'if set, the point in time the process is known to have started'}) 
	started_at: Date;

	/**
	 * if set, the point in time that the process is known to have finished
	 */	
	@ApiProperty({ description: 'if set, the point in time that the process is known to have finished'}) 
	finished_at: Date;

	/**
	 * if true, the process executed successfully
	 */	
	@ApiProperty({ description: 'if true, the process executed successfully'}) 
	succeeded: boolean;

	/**
	 * if true, this row is &quot;logically&quot; deleted. that is, its not to be used any more and is kept for historical/audit purposes only
	 */	
	@ApiProperty({ description: 'if true, this row is &quot;logically&quot; deleted. that is, its not to be used any more and is kept for historical/audit purposes only'}) 
	is_deleted: boolean;

	/**
	 * the id of the user that made the last change or in the case of its initial creation, the user that created it
	 */	
	@ApiProperty({ description: 'the id of the user that made the last change or in the case of its initial creation, the user that created it'}) 
	last_changed_by_id: number;

	/**
	 * the point in time this row was first created regardless of modifications
	 */	
	@ApiProperty({ description: 'the point in time this row was first created regardless of modifications'}) 
	createtime: Date;

	/**
	 * the last point in time this row was modified
	 */	
	@ApiProperty({ description: 'the last point in time this row was modified'}) 
	updatetime: Date;

	/**
	 * This was created from the details of the foreign key FK_BACKGROUND_PROCESSS_LASTUSER_LAST_CHANGED_BY_ID_CHANGES and described as: 
	 */	
	@ApiProperty({ type: () => AllUser, description: 'This was created from the details of the foreign key FK_BACKGROUND_PROCESSS_LASTUSER_LAST_CHANGED_BY_ID_CHANGES and described as: '}) 
	last_changed_by: AllUser;

}